backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.36
greenlet>=3.0.0
asyncpg>=0.30.0
redis>=5.2.0
httpx>=0.28.0
PyJWT[crypto]>=2.12.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
qrcode[pil]>=8.0
user-agents>=2.2.0
slowapi>=0.1.9
gunicorn>=23.0.0
